[IMP] im_livechat, *: introduce PublicLivechatGlobalNotificationHandler
*: website_livechat

Task-2900728
